Sadd deprecation warning fix for redis 4 8 1 (#37)
* Adds rake task variant for eliminating entire applications (#1)

* Adds rake task variant for eliminating entire applications

The unsubscribe rake task, when passed two arguments, does not appear to
actually delete any subscriptions. The exact reason for this is not
obvious, but this is far from the only problem with the task:
- the behavior when an app is provided but a queue name is omitted is
  surprising and potentially destructive
- an application that uses multiple queues would require multiple runs
  of the task to fully decomission

Solve both of these problems with a new code path that unsubscribes a
single application in its entirety. This code path already existed, but
was unreachable as a safeguard against shared applications conflicting
across services.

* Bumps version to 0.13.3

* sadd deprecation fix for redis 4-8-1

* version update

* changelog

---------

Co-authored-by: Drew Long <48696687+aalong-tr@users.noreply.github.com>
Co-authored-by: Shyamsunder Tadori <shyamsundertadori@Shyamsunders-MacBook-Pro-2.local>